store/wizard: extract system slice and reducers into dedicated module
