+++ src/Components/CreateImageWizard/steps/UserGroups/tests/UserGroups.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { screen } from '@testing-library/react';
 
-import { wizardInitialState } from '@/store/slices/wizard';
+import { initialState } from '@/store/slices/wizard';
 import { createUser, renderWithRedux, typeWithWait } from '@/test/testUtils';
 
 import UserGroupsStep from '../index';
@@ -12,7 +12,7 @@ describe('UserGroups Component', () => {
     test('pressing Enter in group name input does not trigger page reload', async () => {
       const { store } = renderWithRedux(<UserGroupsStep />, {
         system: {
-          ...wizardInitialState.system,
+          ...initialState.system,
           groups: [
             {
               name: '',
@@ -34,7 +34,7 @@ describe('UserGroups Component', () => {
     test('pressing Enter in group ID input does not trigger page reload', async () => {
       const { store } = renderWithRedux(<UserGroupsStep />, {
         system: {
-          ...wizardInitialState.system,
+          ...initialState.system,
           groups: [
             {
               name: 'developers',
