+++ src/store/slices/wizard/system/index.ts
@@ -1,5 +1,6 @@
 export * from './constants';
 export * from './selectors';
+export * from './slice';
 export { initialState as systemState } from './state';
 export * from './types';
 export * from './utilities';
